All sizes are now stored in 16.16 fixed point.

+++ Twil/Ui/MarginFwd.hpp
@@ -5,7 +5,7 @@
 namespace Twil {
 namespace Ui {
 
-template<std::uint16_t SpaceX, std::uint16_t SpaceY, typename T>
+template<std::int32_t SpaceX, std::int32_t SpaceY, typename T>
 class MarginT;
 
 }
